requirements.txt
Django==1.10.3
django-bower==5.2.0
django-bootstrap3==8.2.3
whitenoise==2.0.6
Pillow==4.1.1
django-crispy-forms==1.6.0
python-magic==0.4.12
psycopg2==2.7.1
django-contact-form==1.3
django-bootstrap3-datetimepicker==2.2.3
djangoajax
django-ajax
django-datetime-widget==0.9.3
python-dateutil==2.6.0
django-debug-toolbar==1.8
django-rosetta==0.7.13
django-allauth==0.32.0
rules==1.2.1
django-modeltranslation==0.12.1

